requirements.txt
flask
xarray
netCDF4
numpy
numpy-stl
